requirements.txt
numpy
streamlit
matplotlib.pyplot